+++ protocols/gossipsub/src/behaviour.rs
@@ -169,7 +169,7 @@ impl From<MessageAuthenticity> for PublishConfig {
         match authenticity {
             MessageAuthenticity::Signed(keypair) => {
                 let public_key = keypair.public();
-                let key_enc = public_key.clone().into_protobuf_encoding();
+                let key_enc = public_key.to_protobuf_encoding();
                 let key = if key_enc.len() <= 42 {
                     // The public key can be inlined in [`rpc_proto::Message::from`], so we don't include it
                     // specifically in the [`rpc_proto::Message::key`] field.
@@ -181,7 +181,7 @@ impl From<MessageAuthenticity> for PublishConfig {
 
                 PublishConfig::Signing {
                     keypair,
-                    author: public_key.into_peer_id(),
+                    author: public_key.to_peer_id(),
                     inline_key: key,
                 }
             }
@@ -2804,10 +2804,6 @@ where
         )
     }
 
-    fn addresses_of_peer(&mut self, _: &PeerId) -> Vec<Multiaddr> {
-        Vec::new()
-    }
-
     fn inject_connected(&mut self, peer_id: &PeerId) {
         // Ignore connections from blacklisted peers.
         if self.blacklisted_peers.contains(peer_id) {
@@ -3201,6 +3197,9 @@ where
                 }
                 NetworkBehaviourAction::ReportObservedAddr { address, score } => {
                     NetworkBehaviourAction::ReportObservedAddr { address, score }
+                }
+                NetworkBehaviourAction::CloseConnection { peer_id, connection } => {
+                    NetworkBehaviourAction::CloseConnection { peer_id, connection }
                 }
             });
         }